third_party/WebKit/LayoutTests/fast/canvas/drawImage-with-broken-image-expected.txt

This test checks behavior of Canvas::drawImage with a broken source image.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.drawImage(null, 0, 0) threw exception TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object..
PASS ctx.drawImage(null, 0, 0, 20, 20) threw exception TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object..
PASS ctx.drawImage(null, 0, 0, 20, 20, 0, 0, 20, 20) threw exception TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object..
PASS ctx.drawImage(invalidImage, 0, 0) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS ctx.drawImage(invalidImage, 0, 0, 20, 20) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS ctx.drawImage(invalidImage, 0, 0, 20, 20, 0, 0, 20, 20) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS ctx.drawImage(invalidImage, 0, 0, 0, 20) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS ctx.drawImage(invalidImage, 0, 0, 0, 20, 0, 0, 20, 20) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS successfullyParsed is true

TEST COMPLETE
